# docs/Architecture/openapi.yaml
openapi: 3.1.0
info:
  title: 闪卡 App 后端 API
  version: 2.5.0
  description: >
    闪卡 App v2.5 接口契约(机器可读版)。行为契约(状态机、幂等、FSRS 排程、
    错误码语义)见 structure-contract.md;需求依据 PRD v2.5(增量继承 v2.4~v2.1)。
    所有业务接口需携带 Authorization: Bearer(register/login 与探针豁免);写操作需携带
    Idempotency-Key。
    限流:写接口 60 req/min/user、IP 5 req/s,超限返回 429 + Retry-After(详见 structure-contract.md 1.6)。
    5xx(INTERNAL_ERROR / GENERATION_FAILED)由全局中间件统一处理,未逐操作声明。
    列表接口 MVP 暂不分页;幂等重放返回首次成功结果(2xx,含 204)。
    旧 /pdfs 兼容路径保留过渡期,内部委托项目接口同一业务语义(structure-contract 6.2)。
servers:
  - url: https://shanka.kbzz1.top
security:
  - BearerAuth: []
...
paths:
...
  /pdfs:
    post:
      tags: [pdf]
      summary: 上传 PDF,服务端异步解析(文本层 + 目录结构)
      operationId: uploadPdf
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: PDF 文件(HTTPS 传输)
      responses:
        '201':
          description: 已接收,进入 PARSING
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PdfFile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
    get:
      tags: [pdf]
      summary: 最近使用的 PDF 列表
      operationId: listPdfs
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/PdfFile'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /pdfs/{file_id}:
    parameters:
      - $ref: '#/components/parameters/FileId'
    get:
      tags: [pdf]
      summary: PDF 详情与章节列表(解析完成前可轮询)
      operationId: getPdf
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PdfFile'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [pdf]
      summary: 删除 PDF 元数据(可原子取消活跃制卡任务);无项目旧数据仍返回 TASK_IN_PROGRESS
      operationId: deletePdf
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - name: abandon_pre_generation_tasks
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: 同一事务内放弃 DRAFT/样卡阶段任务;GENERATING 仍返回冲突
        - name: cancel_active_tasks
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: 二次确认后同一事务内取消全部活跃制卡任务(含 GENERATING)并继续删除
      responses:
        '204':
          description: 已删除
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
  /pdfs/{file_id}/chapters/{chapter_id}:
    parameters:
      - $ref: '#/components/parameters/FileId'
      - $ref: '#/components/parameters/ChapterId'
    delete:
      tags: [pdf]
      summary: 删除章节(关联 knowledge_points.chapter_id 置 null;历史任务快照不受影响)
      operationId: deleteChapter
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '204':
          description: 已删除
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
    patch:
      tags: [pdf]
      summary: 修改章节名称 / 起始页 / 结束页(PRD 4.1 第 3 步、AC-02)
      operationId: updateChapter
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChapterUpdateRequest'
      responses:
        '200':
          description: 已更新
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chapter'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'

  # ---------- API Key ----------
...
  /decks:
    get:
      tags: [decks]
      summary: 牌组列表(含进度摘要);支持 project_id 归属过滤
      operationId: listDecks
      parameters:
        - name: project_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Deck'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [decks]
      summary: 新建牌组(手动独立牌组 project_id=null)
      operationId: createDeck
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 64
                project_id:
                  type: [string, 'null']
                  format: uuid
                  description: 可选归属项目;缺省/null 为独立牌组
      responses:
        '201':
          description: 已创建
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deck'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /decks/{deck_id}:
    parameters:
      - $ref: '#/components/parameters/DeckId'
    get:
      tags: [decks]
      summary: 牌组详情与进度(真实聚合数据)
      operationId: getDeck
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deck'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [decks]
      summary: 牌组改名(version 递增供缓存刷新)
      operationId: renameDeck
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 64
      responses:
        '200':
          description: 改名后的牌组(含真实进度)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deck'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
    delete:
      tags: [decks]
      summary: 删除牌组及其卡片、复习状态与统计(可原子取消活跃制卡任务)
      operationId: deleteDeck
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - name: abandon_pre_generation_tasks
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: 同一事务内放弃 DRAFT/样卡阶段任务;GENERATING 仍返回冲突
        - name: cancel_active_tasks
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: 二次确认后同一事务内取消全部活跃制卡任务(含 GENERATING)并继续删除
      responses:
        '204':
          description: 已删除
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
  /decks/{deck_id}/deletion-preflight:
    parameters:
      - $ref: '#/components/parameters/DeckId'
    get:
      tags: [decks]
      summary: 删除牌组预检(只读;返回影响范围与活跃任务可执行动作)
      operationId: deckDeletionPreflight
      parameters:
        - name: cancel_active_tasks
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: 预检是否展示“取消活跃任务并删除”动作;不修改数据
      responses:
        '200':
          description: 当前删除影响与阻塞任务
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeletionPreflight'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: opaque Bearer session token(注册/登录签发,服务端只存 SHA-256 摘要;决策 D-05)
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: 幂等键,客户端生成;重复请求返回首次成功结果(见 structure-contract.md 1.3)
    FileId:
      name: file_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    ChapterId:
      name: chapter_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    DeckId:
      name: deck_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    PdfStatus:
      type: string
      enum: [PENDING, PARSING, PARSED, FAILED]
    TaskStatus:
      type: string
      enum: [DRAFT, SAMPLE_GENERATING, AWAITING_SAMPLE_CONFIRMATION, GENERATING, COMPLETED, FAILED, ABANDONED]
      description: V2.5 七态(structure-contract 4.1);历史 PENDING→DRAFT、RUNNING→GENERATING、CANCELLED→ABANDONED、PAUSED→FAILED(LEGACY_PAUSED_TASK)
...
    DeckSource:
      type: string
      enum: [MANUAL, IMPORTED, GENERATED]
      description: GENERATED 为 PDF 制卡新建的归属牌组
...
    Error:
      type: object
      required: [error]
      description: 统一错误响应,外层 error 包装(契约 1.4)
      properties:
        error:
          type: object
          required: [code, message, localization_key]
          properties:
            code:
              type: string
              description: 稳定错误码(见 structure-contract.md 第 7 章)
            message:
              type: string
              description: 面向用户的补充信息,不含内部细节
            localization_key:
              type: string
            actions:
              type: [array, 'null']
              items:
                type: string
              description: 可执行动作提示,如 ABANDON_AND_RETRY / VIEW_TASKS / WAIT_FOR_TERMINAL
            details:
              type: [object, 'null']
              additionalProperties: true
...
    Chapter:
      type: object
      required: [chapter_id, name, start_page, end_page]
      properties:
        chapter_id:
          type: string
          format: uuid
        name:
          type: string
        start_page:
          type: integer
          minimum: 1
        end_page:
          type: integer
          minimum: 1
    ChapterUpdateRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          description: 章节名称,可修改
        start_page:
          type: integer
          minimum: 1
        end_page:
          type: integer
          minimum: 1
      description: 至少提供一个字段;未提供的字段保持不变(PRD 5.2)
    PdfFile:
      type: object
      required: [file_id, filename, size_bytes, status, created_at]
      properties:
        file_id:
          type: string
          format: uuid
        filename:
          type: string
        size_bytes:
          type: integer
        status:
          $ref: '#/components/schemas/PdfStatus'
        error_code:
          type: [string, 'null']
          description: PDF_PARSE_FAILED / PDF_TOC_MISSING
        chapters:
          type: [array, 'null']
          items:
            $ref: '#/components/schemas/Chapter'
        created_at:
          type: string
          format: date-time
...
    Deck:
      type: object
      required:
        - deck_id
        - name
        - source
        - card_count
        - due_count
        - mastered_card_count
        - review_count
        - mastery_ratio
        - created_at
        - updated_at
        - version
      properties:
        deck_id:
          type: string
          format: uuid
        name:
          type: string
        source:
          $ref: '#/components/schemas/DeckSource'
        project_id:
          type: [string, 'null']
          format: uuid
          description: 归属学习项目;null = 手动/独立牌组
        card_count:
          type: integer
        due_count:
          type: integer
        mastered_card_count:
          type: integer
        review_count:
          type: integer
        mastery_ratio:
          type: number
          description: mastered_card_count / card_count,card_count 为 0 时为 0
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        version:
          type: string
          description: 版本号,客户端刷新缓存用
...
    DeletionTaskBlocker:
      type: object
      required: [task_id, status, can_abandon, can_cancel, allowed_actions]
      properties:
        task_id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/TaskStatus'
        internal_stage:
          type: [string, 'null']
          enum: [PLANNING, GENERATING, SCORING, PUBLISHING, null]
        project_id:
          type: [string, 'null']
          format: uuid
        deck_id:
          type: [string, 'null']
          format: uuid
        can_abandon:
          type: boolean
        can_cancel:
          type: boolean
          description: 是否可随资源删除一并取消
        allowed_actions:
          type: array
          items:
            type: string
            enum: [ABANDON_AND_RETRY, CANCEL_AND_DELETE, WAIT_FOR_TERMINAL, VIEW_TASKS]
    DeletionPreflight:
      type: object
      required: [resource_type, resource_id, can_delete, blockers, abandonable_task_ids, has_uncancellable_tasks, cancelable_task_ids, can_cancel, actions, impact]
      properties:
        resource_type:
          type: string
          enum: [PROJECT, DECK]
        resource_id:
          type: string
          format: uuid
        can_delete:
          type: boolean
        blockers:
          type: array
          items:
            $ref: '#/components/schemas/DeletionTaskBlocker'
        abandonable_task_ids:
          type: array
          items:
            type: string
            format: uuid
        has_uncancellable_tasks:
          type: boolean
        cancelable_task_ids:
          type: array
          items:
            type: string
            format: uuid
          description: 可随资源一并取消的活跃制卡任务
        can_cancel:
          type: boolean
          description: 是否允许采用取消任务后删除的决策
        actions:
          type: array
          items:
            type: string
            enum: [ABANDON_AND_RETRY, CANCEL_AND_DELETE, WAIT_FOR_TERMINAL, VIEW_TASKS]
        impact:
          type: object
          additionalProperties: true
...
  responses:
    BadRequest:
      description: 请求非法
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: >-
        受保护接口:缺失、非法、撤销或过期的 Bearer 会话(AUTH_REQUIRED / AUTH_INVALID,携带 WWW-Authenticate: Bearer);
        登录接口:凭据错误(INVALID_CREDENTIALS)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: 状态冲突或幂等内容不一致
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    RateLimited:
      description: 限流
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'